requirements.txt
alabaster==0.7.12
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.5.0
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
Cheetah3==3.2.6
click==8.0.3
click-plugins==1.1.1
click-threading==0.5.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
duplicity==0.8.21
fasteners==0.14.1
Flask==2.0.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gps==3.22
gvm-tools==24.8.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httplib2==0.20.2
hyperframe==6.0.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.6.4
ipython==7.31.1
itsdangerous==2.1.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpath-ng==1.7.0
jsonschema==3.2.0
kaitaistruct==0.9
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.8.1
livereload==2.6.3
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mitmproxy==6.0.2
mkdocs==1.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
netifaces==0.11.0
networkx==2.4
nftables==0.1
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
parso==0.8.1
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.0.0
pipx==1.0.0
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
publicsuffix2==2.20191221
py==1.10.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
pygccxml==2.2.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt-Qwt==1.2.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-gvm==24.8.0
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.8
requests==2.25.1
robotframework==7.0.1
robotframework-jsonlibrary==0.5
robotframework-sshlibrary==3.8.0
roman==3.3
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scapy==2.4.4
scipy==1.8.0
scp==0.15.0
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.1.0
soupsieve==2.3.1
Sphinx==4.3.2
ssh-import-id==5.11
sympy==1.9
systemd-python==234
thrift==0.16.0
tornado==6.1
traitlets==5.1.1
ttkbootstrap==1.10.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
urwid==2.1.2
usb-creator==0.3.7
userpath==1.8.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wsproto==1.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
